Adjusted remaining identifying DB columns to use big integer fields. Added 'User' DB seeder with default user assignment to 'admin-root' role. Added 'Plan' DB seeder.

+++ database/migrations/2018_01_29_140758_create_confirmation_tokens_table.php
@@ -14,7 +14,7 @@ class CreateConfirmationTokensTable extends Migration {
         Schema::create('confirmation_tokens', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table
-                ->integer('user_id')
+                ->bigInteger('user_id')
                 ->unsigned()
                 ->index();
             $table->string('token');
